requirements.txt
aiogram==3.7.0
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.1
